+++ src/main/java/lotto/domain/purchase/lottocount/CashMapper.java
@@ -1,7 +1,7 @@
 package lotto.domain.purchase.lottocount;
 
 public final class CashMapper {
-    public static Integer mapToInteger(String cashInput) {
-        return Integer.parseInt(cashInput);
+    public static Long mapToInteger(String cashInput) {
+        return Long.parseLong(cashInput);
     }
 }
