+++ src/main/java/lotto/domain/purchase/lottocount/LotteryCountService.java
@@ -5,7 +5,7 @@ import lotto.domain.view.InputView;
 import lotto.domain.view.OutputView;
 
 public class LotteryCountService {
-    public Integer getCash() {
+    public Long getCash() {
         // 금액 입력 뷰
         InputView.getCash();
         // 금액 입력받기
@@ -17,9 +17,9 @@ public class LotteryCountService {
     /***
      * {@link LottoCountCalculator}에게 티켓 개수 구하기를 위임
      */
-    public Integer getTicketCount(Integer cash) {
+    public Integer getTicketCount(Long cash) {
         // 금액에 대한 티켓 개수 계산
-        Integer ticketCount = LottoCountCalculator.divideByLottoPrice(cash);
+        int ticketCount = LottoCountCalculator.divideByLottoPrice(cash);
         // 티켓 개수 출력 뷰
         OutputView.printLottoAmount(ticketCount);
         return ticketCount;
